improve error string if parse_binds is not set
When using rtic-syntax as a library and parsing the below

    let (_app, _analysis) = crate::parse2(
        quote!(),
        quote!(
            mod app {
                #[task(binds = UART0)]
                fn foo(_: foo::Context) {}
            }
        ),
        Settings::default(),
    )
    .unwrap();

the user it met with a cryptic Error("unexpected argument"). What is not
obvious is that parse_binds in the Settings struct must be set if binds
are to be parsed.

This commit adds an improved error and tests for it. Resolves #46.